+++ cmd/srcode/srcode_test.go
@@ -1,6 +1,7 @@
 package main
 
 import (
+	"errors"
 	"fmt"
 	"github.com/creekorful/srcode/internal/codebase"
 	"github.com/creekorful/srcode/internal/codebase_mock"
@@ -321,7 +322,7 @@ func TestPwd(t *testing.T) {
 	}
 }
 
-func TestRunCmd(t *testing.T) {
+func TestRunScript(t *testing.T) {
 	mockCtrl := gomock.NewController(t)
 	defer mockCtrl.Finish()
 
@@ -455,15 +456,14 @@ func TestBulkGit(t *testing.T) {
 	codebaseProviderMock.EXPECT().Open(cwd).Return(codebaseMock, nil)
 
 	codebaseMock.EXPECT().
-		BulkGIT([]string{"pull", "--rebase", "--prune"}, gomock.Any()).
-		Do(func(args []string, ch chan<- string) { close(ch) })
+		BulkGIT([]string{"pull", "--rebase", "--prune"}, b)
 
 	if err := app.getCliApp().Run([]string{"srcode", "bulk-git", "pull", "--rebase", "--prune"}); err != nil {
 		t.Fail()
 	}
 }
 
-func TestSetCmd(t *testing.T) {
+func TestScript(t *testing.T) {
 	mockCtrl := gomock.NewController(t)
 	defer mockCtrl.Finish()
 
@@ -479,28 +479,38 @@ func TestSetCmd(t *testing.T) {
 	}
 
 	// test with no enough args should fails
-	if err := app.getCliApp().Run([]string{"srcode", "set-cmd"}); err != errWrongSetCmdUsage {
-		t.Errorf("got %v want %v", err, errWrongSetCmdUsage)
-	}
-	if err := app.getCliApp().Run([]string{"srcode", "set-cmd", "test"}); err != errWrongSetCmdUsage {
-		t.Errorf("got %v want %v", err, errWrongSetCmdUsage)
+	if err := app.getCliApp().Run([]string{"srcode", "script"}); err != errWrongScriptUsage {
+		t.Errorf("got %v want %v", err, errWrongScriptUsage)
 	}
 
 	codebaseMock := codebase_mock.NewMockCodebase(mockCtrl)
 
+	// test no project in current directory
+	codebaseProviderMock.EXPECT().Open(cwd).Return(codebaseMock, nil)
+	codebaseMock.EXPECT().Projects().Return(map[string]codebase.ProjectEntry{}, nil)
+	codebaseMock.EXPECT().LocalPath().Return("test-12")
+
+	if err := app.getCliApp().Run([]string{"srcode", "script", "test", "@go-test"}); !errors.Is(err, manifest.ErrNoProjectFound) {
+		t.Fail()
+	}
+
 	// test set local command
 	codebaseProviderMock.EXPECT().Open(cwd).Return(codebaseMock, nil)
-	codebaseMock.EXPECT().SetCommand("test", "@go-test", false)
+	codebaseMock.EXPECT().Projects().Return(map[string]codebase.ProjectEntry{"test-12": {}}, nil)
+	codebaseMock.EXPECT().LocalPath().Return("test-12")
+	codebaseMock.EXPECT().SetScript("test", []string{"@go-test"}, false)
 
-	if err := app.getCliApp().Run([]string{"srcode", "set-cmd", "test", "@go-test"}); err != nil {
+	if err := app.getCliApp().Run([]string{"srcode", "script", "test", "@go-test"}); err != nil {
 		t.Fail()
 	}
 
 	// test set global command
 	codebaseProviderMock.EXPECT().Open(cwd).Return(codebaseMock, nil)
-	codebaseMock.EXPECT().SetCommand("go-test", "go test -race -v ./...", true)
+	codebaseMock.EXPECT().Projects().Return(map[string]codebase.ProjectEntry{"test-42": {}}, nil)
+	codebaseMock.EXPECT().LocalPath().Return("test-42")
+	codebaseMock.EXPECT().SetScript("go-test", []string{"go test -race -v ./..."}, true)
 
-	if err := app.getCliApp().Run([]string{"srcode", "set-cmd", "--global", "go-test", "go", "test", "-race", "-v", "./..."}); err != nil {
+	if err := app.getCliApp().Run([]string{"srcode", "script", "--global", "go-test", "go", "test", "-race", "-v", "./..."}); err != nil {
 		t.Fail()
 	}
 }
@@ -583,6 +593,42 @@ func TestRmProject(t *testing.T) {
 	}
 
 	if b.String() != "Successfully deleted Contributing/Test\n" {
+		t.Fail()
+	}
+}
+
+func TestHook(t *testing.T) {
+	mockCtrl := gomock.NewController(t)
+	defer mockCtrl.Finish()
+
+	codebaseProviderMock := codebase_mock.NewMockProvider(mockCtrl)
+
+	b := &strings.Builder{}
+
+	app := app{
+		codebaseProvider: codebaseProviderMock,
+		writer:           b,
+	}
+
+	cwd, err := os.Getwd()
+	if err != nil {
+		t.FailNow()
+	}
+
+	// test with no enough args should fails
+	if err := app.getCliApp().Run([]string{"srcode", "hook"}); err != errWrongHookUsage {
+		t.Errorf("got %v want %v", err, errWrongHookUsage)
+	}
+
+	codebaseMock := codebase_mock.NewMockCodebase(mockCtrl)
+	codebaseProviderMock.EXPECT().Open(cwd).Return(codebaseMock, nil)
+	codebaseMock.EXPECT().SetHook("lint").Return(nil)
+	codebaseMock.EXPECT().LocalPath().Return("Contributing/Test")
+	if err := app.getCliApp().Run([]string{"srcode", "hook", "lint"}); err != nil {
+		t.Fail()
+	}
+
+	if b.String() != "Successfully applied hook `lint` to /Contributing/Test\n" {
 		t.Fail()
 	}
 }