+++ internal/repository/repository.go
@@ -2,6 +2,7 @@ package repository
 
 import (
 	"github.com/creekorful/srcode/internal/cmd"
+	"io"
 	"os/exec"
 )
 
@@ -16,7 +17,7 @@ type Repository interface {
 	Remote(name string) (string, error)
 	Config(key string) (string, error)
 	SetConfig(key, value string) error
-	RawCmd(args []string) (string, error)
+	RawCmd(args []string, writer io.Writer) error
 	Head() (string, error)
 	IsDirty() (bool, error)
 }
@@ -65,8 +66,13 @@ func (gwr *gitWrapperRepository) SetConfig(key, value string) error {
 	return err
 }
 
-func (gwr *gitWrapperRepository) RawCmd(args []string) (string, error) {
-	return gwr.execWithOutput(args...)
+func (gwr *gitWrapperRepository) RawCmd(args []string, writer io.Writer) error {
+	command := exec.Command("git", args...)
+	command.Dir = gwr.path
+	command.Stdout = writer
+	command.Stderr = writer
+
+	return command.Run()
 }
 
 func (gwr *gitWrapperRepository) Head() (string, error) {