+++ CHANGELOG.md
@@ -7,6 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.7.0] - 2021-02-10
+
+## Changed
+
+- Improve output of bulk-git.
+- [#18](https://github.com/creekorful/srcode/issues/18) set-cmd: edit with $EDITOR if command not directly provided.
+- set-cmd: allow edition of existing command.
+- set-cmd: rename to script.
+- [#19](https://github.com/creekorful/srcode/issues/19) script: now we are making script and not just command.
+- [#20](https://github.com/creekorful/srcode/issues/20) add git hook support.
+
 ## [0.6.0] - 2021-02-09
 
 ## Added
@@ -70,7 +81,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Initial pre-release.
 
-[Unreleased]: https://github.com/creekorful/srcode/compare/v0.6.0...HEAD
+[Unreleased]: https://github.com/creekorful/srcode/compare/v0.7.0...HEAD
+[0.7.0]: https://github.com/creekorful/srcode/compare/v0.7.0...HEAD
 [0.6.0]: https://github.com/creekorful/srcode/compare/v0.6.0...HEAD
 [0.5.0]: https://github.com/creekorful/srcode/compare/v0.5.0...HEAD
 [0.5.0]: https://github.com/creekorful/srcode/compare/v0.5.0...HEAD